Complete Etch A Sketch Project - First Edition
The html file was set up with all of the elements displayed except the
colored divs where the coloring is done by the user.

The css file contains many of the types that will later be applied to
the created divs.

The js file starts off initializing all of the query selectors and
adding in some of the event listeners with more basic functions. The
latter half is where the divs are created and colors are added to each
when hovered over.